+++ src/App.jsx
@@ -2,9 +2,12 @@ import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
+import Game from './models/game'
 
 function App() {
   const [count, setCount] = useState(0)
+  let game = new Game(1);
+  game.startGame();
 
   return (
     <>